+++ src/cli/_common.py
@@ -2,11 +2,14 @@
 Common code for CLI modules.
 """
 
+import sys
 from ipaddress import IPv4Address
 from pathlib import Path
 from typing import Dict, Set
 
+import click
 import docker
+from docker.client import DockerClient
 from docker.models.containers import Container
 
 from dcos_e2e.cluster import Cluster
@@ -38,11 +41,23 @@ WORKSPACE_DIR_LABEL_KEY = 'dcos_e2e.workspace_dir'
 VARIANT_LABEL_KEY = 'dcos_e2e.variant'
 
 
+def docker_client() -> DockerClient:
+    """
+    Return a Docker client.
+    """
+    try:
+        return docker.from_env(version='auto')
+    except docker.errors.DockerException:
+        message = 'Error: Cannot connect to Docker.'
+        click.echo(message, err=True)
+        sys.exit(1)
+
+
 def existing_cluster_ids() -> Set[str]:
     """
     Return the IDs of existing clusters.
     """
-    client = docker.from_env(version='auto')
+    client = docker_client()
     filters = {'label': CLUSTER_ID_LABEL_KEY}
     containers = client.containers.list(filters=filters)
     return set(
@@ -103,7 +118,7 @@ class ClusterContainers:
         """
         Return all containers in this cluster of a particular node type.
         """
-        client = docker.from_env(version='auto')
+        client = docker_client()
         filters = {
             'label': [
                 self._cluster_id_label,